+++ packages/javascript/lib/verify/isEmail.ts
@@ -1,6 +1,11 @@
-export function isEmail(email: string): boolean {
+export function isEmail(email: string, onlyLowerCase = false): boolean {
 	// RFC2822 Email Validation
-	return /^[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$/.test(
-		email
+	const char = onlyLowerCase ? 'a-z' : 'a-zA-Z';
+
+	const regex = new RegExp(
+		`^[${char}0-9!#$%&'*+/=?^_\`{|}~-]+(?:\\.[${char}0-9!#$%&'*+/=?^_\`{|}~-]+)*@(?:[${char}0-9](?:[${char}0-9-]*[${char}0-9])?\\.)+[${char}0-9](?:[${char}0-9-]*[${char}0-9])?$`,
+		'g'
 	);
+
+	return regex.test(email);
 }