+++ packages/dart/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog (Dart)
 
+## 1.1.6 (2025--)
+
+- `isEmail`: add `onlyLowerCase` parameter
+
 ## 1.1.5 (2025-03-06)
 
 - Update `README.md`